Screen: Add pageflip diagnostics for Intel gpu's.
This is naturally limited to Linux and OSX(with kernel driver loaded
and the dangerous override for Intel set), and the register mapping is
chosen based on IVB and so far only tested for IvyBridge IGP. May or
may not work on earlier or later gens, too lazy to look it up atm.

-> This to debug the latest Apple provided macOS disasters for 10.14
Mojave.