Screen/OSX: Bring CGLSetFullScreenOnDisplay() back from the dead.
All the testing with the new Intel gpu pageflip diagnostic on 10.13.6
and
10.14.0 with Intel Ironlake and Intel IvyBridge gfx, as well as testing
on
old Nvidia’s of the GT330M and GT120M series and with a brand-new
Radeon Pro WX4100 showed that the root cause of most of the timing
misery at least on 10.13 and 10.14 seems to be compositor interference.
We are not in control of page flipping, but all our fullscreen windows
get redirected to the compositor, which then screws up all our timing
and timestamping, as not only PTB’s internal diagnostic shows, but also
measurement with a Datapixx.
Given that the idiots apparently totally screwed up their Cocoa-based
fullscreen windowing support, which was supposed to be optimised, but
apparently isn’t, i made a bold move: What about doing the exact
opposite of everything Apple recommends wrt. performant graphics
programming? Use the old CGLSetFullScreenOnDisplay() call again to
switch the OpenGL rendering context into fullscreen mode on a display,
completely bypassing the Cocoa API pile of shit. This in the faint hope
that the way we did it over 4 years ago, before
CGLSetFullScreenOnDisplay() got deprecated in 10.7 with the strong
recommendation by Apple to avoid it, is now no longer the most broken
way. Apparently the asshats managed to screw up their shiny new stuff
so badly that the old buggy solution from years ago is now the less
buggy solution.

And voila! Going back to CGLSetFullScreenOnDisplay() for fullscreen
windows immediately resolved all(!) the catastrophic visual timing and
timestamping problems under 10.14 Mojave with Intel gfx. At least as
tested on single display stimulation on single- and dual-display
setups. Lets hope these results will hold up for other OS + GPUs +
display-config combos as well.

-> Promising backup commit.