Screen/Linux: Implement homegrown P-Sync Flip support.
P-Sync for "Poor man's adaptive sync" is our own
homegrown pure software trickery implementation of
a Flip mechanism similar to AMD FreeSync or NVidia
G-Sync, which allows to achieve stimulus onset not only
at multiples of the video refresh duration, but also
in between.

Working principle:

Wait until 1 video refresh duration before the refresh
cycle in which the target stimulus onset time 'tWhen' is
located. Then MMIO program the display engine to stop
scanout at the end of the current refresh cycle, inside
the VBlank before the target time tWhen and execute the
bufferswap/page-flip, to at the end of the refresh cycle,
the engine is stopped and ready for instant restart, with
the proper new visual stimulus set as scanout buffer. Then
we wait until the target time tWhen and then MMIO program
the display engine to restart scanout, starting a fresh
scanout cycle with our new stimulus as closely to tWhen
as possible. We use our own beamposition timestamping to
compute the end-time of the vblank in which the engine was
restarted, and therefore stimulus onset time.

This only works on Linux + a native X-Server under X11/GLX
with the open-source DRM/KMS graphics/display drivers with
AMD gpu's that have at least a DCE-4 display engine, ie.
the HD-5000 series Evergreen and later up to Polaris DCE-11,
maybe also with Vega DCE-12. It requires a display device
that is tolerant of multi-millisecond dropouts in the video
stream - up to 1 video refresh duration. So far, this was
only somewhat successfully tested on a single DisplayPort
driven ASUS ROG PG "NVidia G-Sync" compatible monitor. The
method failed on other non-G-Sync displays connected via
DisplayPort, HDMI, and DVI-D (the displays blanked the whole
time) and also on an analog VGA CRT, which didn't blank, but
got a distorted or jumping visual stimulus. In other words,
sadly this may be of very limited use on only a very few
display devices.

We use as little of the OS OML sync control api as possible
for our trick, because the OS vblank counter and vblank
timestamping is compromised immediatly after a p-Sync style
flip for at least one, possibly two video refresh cycles.
Therefore we can't trust the api's to provide sensible
data or behavior. We use time-based scheduling instead of
msc based scheduling, and reimplement beampos vblank
timestamping in a helper function, so we can get precise
vblank start/end time whenever we need, instantaneously
without use of temporarily compromised OS facilities.

This should allow reducing the dead-time to less than
two video refresh cycles - down to almost one cycle. In
practice the only one display that kind of works actually
blanks for such < 2 refresh durations.

Datapixx timestamping showed that stimulus onset was pretty
accurate wrt. tWhen and timestamps were correct, but no
measurement with photo-diodes or such was done to actually
verify correct stimulus onset timing on the only G-Sync
display that visually seemed to work.

-> Not sure if this homegrown method has any real practical
   value, but maybe worth merging on the off-chance it is
   good for something?